+++ Rain/src/com/swiftrunner/rain/entity/projectile/Projectile.java
@@ -8,7 +8,7 @@ import com.swiftrunner.rain.graphics.Sprite;
 
 public abstract class Projectile extends Entity {
 	
-	protected final int xOrigin, yOrigin;
+	protected final double xOrigin, yOrigin;
 	protected int distance;
 	protected double angle;
 	protected double x, y;
@@ -19,7 +19,7 @@ public abstract class Projectile extends Entity {
 	
 	protected final Random random = new Random();
 	
-	public Projectile(int x, int y, double dir) {
+	public Projectile(double x, double y, double dir) {
 		xOrigin = x;
 		yOrigin = y;
 		angle = dir;
