+++ Rain/src/com/swiftrunner/rain/level/Level.java
@@ -152,15 +152,15 @@ public class Level {
 	
 	public List<Entity> getEntities(Entity e, int radius){
 		List<Entity> result = new ArrayList<Entity>();
-		int ex = e.getX();
-		int ey = e.getY();
+		int ex = (int)e.getX();
+		int ey = (int)e.getY();
 		
 		for(int i=0; i<entities.size(); i++){
 			Entity entity = entities.get(i);
-			int x = entity.getX();
-			int y = entity.getY();
-			int dx = Math.abs(x - ex);
-			int dy = Math.abs(y - ey);
+			int x = (int)entity.getX();
+			int y = (int)entity.getY();
+			double dx = Math.abs(x - ex);
+			double dy = Math.abs(y - ey);
 			double distance = Math.sqrt((dx * dx) + (dy * dy));
 			
 			if(distance <= radius){ result.add(entity); }
@@ -171,15 +171,15 @@ public class Level {
 	
 	public List<Player> getPlayers(Entity e, int radius){
 		List<Player> result = new ArrayList<Player>();
-		int ex = e.getX();
-		int ey = e.getY();
+		int ex = (int)e.getX();
+		int ey = (int)e.getY();
 		
 		for(int i=0; i<players.size(); i++){
 			Player player = players.get(i);
-			int x = player.getX();
-			int y = player.getY();
-			int dx = Math.abs(x - ex);
-			int dy = Math.abs(y - ey);
+			int x = (int)player.getX();
+			int y = (int)player.getY();
+			double dx = Math.abs(x - ex);
+			double dy = Math.abs(y - ey);
 			double distance = Math.sqrt((dx * dx) + (dy * dy));
 			if(distance <= radius){ result.add(player); }
 		}
